requirements/base.txt
Django==1.8
djangorestframework==3.1.1



